extends SimpleFunctionRector to SimpleFilter as well
