+++ config/level/twig/twig112.yml
@@ -4,3 +4,5 @@ services:
         $oldToNewClasses:
             Twig_Function_Node: 'Twig_SimpleFunction'
             Twig_Function: 'Twig_SimpleFunction'
+            Twig_Filter: 'Twig_SimpleFilter'
+            Twig_Test: 'Twig_SimpleTest'
