+++ config/level/twig/twig20.yml
@@ -5,3 +5,5 @@ services:
             Twig_SimpleFilter: 'Twig_Filter'
             # see https://twig.symfony.com/doc/1.x/deprecated.html#functions
             Twig_SimpleFunction: 'Twig_Function'
+            # see https://github.com/bolt/bolt/pull/6596
+            Twig_SimpleTest: 'Twig_Test'
