+++ packages/Twig/tests/Rector/SimpleFunctionAndFilterRector/Source/TwigFunctionMethod.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Rector\Twig\Tests\Rector\SimpleFunctionAndFilterRector\Source;
+
+final class TwigFunctionMethod
+{
+
+}
