--- packages/Twig/tests/Rector/SimpleFunctionRector/Source/TwigSimpleFunction.php
@@ -1,8 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Rector\Twig\Tests\Rector\SimpleFunctionRector\Source;
-
-final class TwigSimpleFunction
-{
-
-}
